# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server server.",
        "title": "APIs",
        "termsOfService": "https://www.aofiee.dev/",
        "contact": {
            "name": "API Support",
            "url": "https://www.aofiee.dev/",
            "email": "aofiee@aofiee.dev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9089",
    "basePath": "/",
    "paths": {
        "/v1/api/todo": {
            "get": {
                "description": "Delete todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODO"
                ],
                "summary": "Delete todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "order_by",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "asc",
                        "name": "asc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "description",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODO"
                ],
                "summary": "Update todo",
                "parameters": [
                    {
                        "description": "UpdateTodo",
                        "name": "UpdateTodo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.TodoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODO"
                ],
                "summary": "Create todo",
                "parameters": [
                    {
                        "description": "CreateTodo",
                        "name": "CreateTodo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.TodoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/api/todo/{id}": {
            "delete": {
                "description": "Delete todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODO"
                ],
                "summary": "Delete todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/webhook/line": {
            "post": {
                "description": "Handles webhook events from LINE Messaging API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LINE"
                ],
                "summary": "LINE Webhook",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.TodoRequest": {
            "type": "object",
            "required": [
                "date",
                "status",
                "title"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        "IN_PROGRESS",
                        "COMPLETE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/http.TodoStatus"
                        }
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "http.TodoStatus": {
            "type": "string",
            "enum": [
                "IN_PROGRESS",
                "COMPLETE"
            ],
            "x-enum-varnames": [
                "TodoStatusInProgress",
                "TodoStatusComplete"
            ]
        }
    }
}